+++ src/main/java/com/luckyvicky/woosan/domain/board/service/AIService.java
@@ -1,0 +1,11 @@
+package com.luckyvicky.woosan.domain.board.service;
+
+import com.cybozu.labs.langdetect.LangDetectException;
+import com.luckyvicky.woosan.domain.board.dto.BoardDTO;
+
+import java.io.IOException;
+
+public interface AIService {
+    BoardDTO translateBoardDetailPage(BoardDTO boardDTO) throws IOException;
+    String summaryBoardDetailPage(BoardDTO boardDTO) throws IOException, LangDetectException;
+}
