+++ src/main/java/com/luckyvicky/woosan/domain/board/service/SummaryService.java
@@ -1,10 +1,10 @@
-package com.luckyvicky.woosan.domain.board.service;
-
-import com.cybozu.labs.langdetect.LangDetectException;
-import com.luckyvicky.woosan.domain.board.dto.BoardDTO;
-
-import java.io.IOException;
-
-public interface SummaryService {
-    String summaryBoardDetailPage(BoardDTO boardDTO) throws IOException, LangDetectException;
-}
+//package com.luckyvicky.woosan.domain.board.service;
+//
+//import com.cybozu.labs.langdetect.LangDetectException;
+//import com.luckyvicky.woosan.domain.board.dto.BoardDTO;
+//
+//import java.io.IOException;
+//
+//public interface SummaryService {
+//    String summaryBoardDetailPage(BoardDTO boardDTO) throws IOException, LangDetectException;
+//}
